requirements.txt
python-dotenv~=1.0.1
pandas~=2.2.3
numpy~=2.2.2
pymongo~=4.11.1
certifi~=2025.1.31
pymongo[srv]
scikit-learn~=1.6.1
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart
setuptools~=75.8.0
PyYAML~=6.0.2
dill~=0.3.9
scipy~=1.15.1

-e .
